refactor: create DemoDialog generic component
